Make code more OOP
