+++ src/main/java/uncleroger/parser/Parser.java
@@ -1,0 +1,218 @@
+package uncleroger.parser;
+
+import uncleroger.exception.AlreadyMarkedException;
+import uncleroger.exception.AlreadyUnmarkedException;
+import uncleroger.exception.EmptyListException;
+import uncleroger.exception.InvalidCommandException;
+import uncleroger.exception.InvalidDeadlineException;
+import uncleroger.exception.InvalidEventFieldOrderException;
+import uncleroger.exception.MissingEventFieldsException;
+import uncleroger.exception.NoDescriptionException;
+import uncleroger.exception.NoEntryYetException;
+import uncleroger.exception.NonPositiveIndexException;
+import uncleroger.task.Deadline;
+import uncleroger.task.Event;
+import uncleroger.task.Todo;
+import uncleroger.ui.TextUi;
+
+import java.util.Arrays;
+
+import static uncleroger.UncleRoger.tasks;
+
+public class Parser {
+
+    private static void handleMarkCommand(String[] words)
+            throws NoEntryYetException, AlreadyMarkedException,
+            NonPositiveIndexException {
+        int userEnteredEntry = Integer.parseInt(words[1]);
+        int userEnteredArrIndex = userEnteredEntry - 1;
+        if (userEnteredEntry <= 0) {
+            throw new NonPositiveIndexException();
+        } else if (userEnteredArrIndex >= tasks.size()) {
+            throw new NoEntryYetException();
+        } else if (tasks.get(userEnteredArrIndex).getIsDone()) {
+            throw new AlreadyMarkedException();
+        } else {
+            tasks.get(userEnteredArrIndex).setIsDone(true);
+            TextUi.printTaskStatus(tasks.get(userEnteredArrIndex));
+        }
+    }
+
+    private static void handleUnmarkedCommand(String[] words)
+            throws NoEntryYetException, AlreadyUnmarkedException, NonPositiveIndexException {
+        int userEnteredEntry = Integer.parseInt(words[1]);
+        int userEnteredArrIndex = userEnteredEntry - 1;
+        if (userEnteredEntry <= 0) {
+            throw new NonPositiveIndexException();
+        } else if (userEnteredArrIndex >= tasks.size()) {
+            throw new NoEntryYetException();
+        } else if (!tasks.get(userEnteredArrIndex).getIsDone()) {
+            throw new AlreadyUnmarkedException();
+        } else {
+            tasks.get(userEnteredArrIndex).setIsDone(false);
+            TextUi.printTaskStatus(tasks.get(userEnteredArrIndex));
+        }
+    }
+
+    private static void handleDeleteCommand(String[] words)
+            throws NonPositiveIndexException, NoEntryYetException {
+        int userEnteredEntry = Integer.parseInt(words[1]);
+        int userEnteredArrIndex = userEnteredEntry - 1;
+        if (userEnteredEntry <= 0) {
+            throw new NonPositiveIndexException();
+        } else if (userEnteredArrIndex >= tasks.size()) {
+            throw new NoEntryYetException();
+        } else {
+            TextUi.printTaskRemove(tasks.get(userEnteredArrIndex));
+            tasks.remove(userEnteredArrIndex);
+        }
+    }
+
+    private static void handleListCommand() throws EmptyListException {
+        if (tasks.isEmpty()) {
+            throw new EmptyListException();
+        } else {
+            TextUi.printList();
+        }
+    }
+
+    private static void handleTodoCommand(String[] words)
+            throws NoDescriptionException {
+        if (words.length < 2) {
+            throw new NoDescriptionException();
+        }
+        String description = combineWordsToSentence(words, 1, words.length);
+        tasks.add(new Todo(description));
+        TextUi.printTaskEntry();
+    }
+
+    private static void handleDeadlineCommand(String[] words)
+            throws NoDescriptionException, InvalidDeadlineException {
+        if (words.length < 2) {
+            throw new NoDescriptionException();
+        }
+        int indexOfBy = -1;
+        for (int i = 0; i < words.length; i++) {
+            if (words[i].equalsIgnoreCase("/by")) {
+                indexOfBy = i;
+                break;
+            }
+        }
+        if (indexOfBy == -1) {
+            throw new InvalidDeadlineException();
+        }
+        String description = combineWordsToSentence(words, 1, indexOfBy);
+        String by = combineWordsToSentence(words, indexOfBy + 1, words.length);
+        tasks.add(new Deadline(by, description));
+        TextUi.printTaskEntry();
+    }
+
+    private static void handleEventCommand(String[] words)
+            throws NoDescriptionException, MissingEventFieldsException,
+            InvalidEventFieldOrderException {
+        if (words.length < 2) {
+            throw new NoDescriptionException();
+        }
+        int indexOfFrom = -1;
+        int indexOfTo = -1;
+        for (int i = 0; i < words.length; i++) {
+            if (words[i].equalsIgnoreCase("/from")) {
+                indexOfFrom = i;
+            }
+            if (words[i].equalsIgnoreCase("/to")) {
+                indexOfTo = i;
+            }
+        }
+        if (indexOfFrom == -1 || indexOfTo == -1) {
+            throw new MissingEventFieldsException();
+        }
+        if (indexOfFrom > indexOfTo) {
+            throw new InvalidEventFieldOrderException();
+        }
+        String description = combineWordsToSentence(words, 1, indexOfFrom);
+        String from = combineWordsToSentence(words, indexOfFrom + 1, indexOfTo);
+        String to = combineWordsToSentence(words, indexOfTo + 1, words.length);
+        tasks.add(new Event(from, to, description));
+        TextUi.printTaskEntry();
+    }
+
+    private static String combineWordsToSentence(String[] words, int start, int end) {
+        words = Arrays.copyOfRange(words, start, end);
+        return String.join(" ", words);
+    }
+
+    public static void parseUserSentence(String sentence)
+            throws InvalidCommandException {
+        String[] words = sentence.split(" ");
+        String command = words[0];
+        switch (command.toLowerCase()) {
+        case "list":
+            try {
+                handleListCommand();
+            } catch (EmptyListException e) {
+                TextUi.printEmptyList();
+            }
+            break;
+        case "mark":
+            try {
+                handleMarkCommand(words);
+            } catch (NoEntryYetException e) {
+                TextUi.printNoEntryYet();
+            } catch (AlreadyMarkedException e) {
+                TextUi.printAlreadyMarked();
+            } catch (NonPositiveIndexException e) {
+                TextUi.printNonPositiveIndex();
+            }
+            break;
+        case "unmark":
+            try {
+                handleUnmarkedCommand(words);
+            } catch (NoEntryYetException e) {
+                TextUi.printNoEntryYet();
+            } catch (AlreadyUnmarkedException e) {
+                TextUi.printAlreadyUnmarked();
+            } catch (NonPositiveIndexException e) {
+                TextUi.printNonPositiveIndex();
+            }
+            break;
+        case "todo":
+            try {
+                handleTodoCommand(words);
+            } catch (NoDescriptionException e) {
+                TextUi.printNoDescription();
+            }
+            break;
+        case "deadline":
+            try {
+                handleDeadlineCommand(words);
+            } catch (NoDescriptionException e) {
+                TextUi.printNoDescription();
+            } catch (InvalidDeadlineException e) {
+                TextUi.printInvalidDeadline();
+            }
+            break;
+        case "event":
+            try {
+                handleEventCommand(words);
+            } catch (NoDescriptionException e) {
+                TextUi.printNoDescription();
+            } catch (MissingEventFieldsException e) {
+                TextUi.printMissingEventFields();
+            } catch (InvalidEventFieldOrderException e) {
+                TextUi.printInvalidEventFieldOrder();
+            }
+            break;
+        case "delete":
+            try {
+                handleDeleteCommand(words);
+            } catch (NonPositiveIndexException e) {
+                TextUi.printNonPositiveIndex();
+            } catch (NoEntryYetException e) {
+                TextUi.printNoEntryYet();
+            }
+            break;
+        default:
+            throw new InvalidCommandException();
+        }
+    }
+}
